scraper/requirements.txt
requests>=2.31.0
python-dotenv>=0.21.0
tqdm>=4.65.0 